feat : 특정 유저의 상 벌점 조회 컨트롤러 추가
- 특정 유저의 상 벌점 조회 컨트롤러를 추가했습니다.
