+++ waldreg/application-layer/acceptance/src/test/java/org/waldreg/acceptance/reward/RewardAcceptanceTest.java
@@ -25,6 +25,8 @@ import org.waldreg.acceptance.user.UserAcceptanceTestHelper;
 import org.waldreg.auth.request.AuthTokenRequest;
 import org.waldreg.controller.reward.tag.request.RewardTagRequest;
 import org.waldreg.controller.reward.tag.response.RewardTagResponse;
+import org.waldreg.controller.reward.users.response.RewardTagWrapperResponse;
+import org.waldreg.controller.reward.users.response.UsersRewardTagResponse;
 import org.waldreg.controller.user.request.UserRequest;
 import org.waldreg.controller.user.response.UserResponse;
 
@@ -918,193 +920,6 @@ public class RewardAcceptanceTest{
     private UsersRewardTagResponse getUsersRewardTagResponse(ResultActions resultActions) throws Exception{
         String content = resultActions.andReturn().getResponse().getContentAsString();
         return objectMapper.readValue(content, UsersRewardTagResponse.class);
-    }
-
-    public final static class UsersRewardTagResponse{
-
-        private final int id;
-        private final String name;
-
-        @JsonProperty("user_id")
-        private final String userId;
-        private final int reward;
-
-        @JsonProperty("reward_infos")
-        private final List<RewardTagWrapperResponse> rewardInfoList;
-
-        private UsersRewardTagResponse(){
-            throw new UnsupportedOperationException("Can not invoke constructor \"UsersRewardTagResponse()\"");
-        }
-
-        private UsersRewardTagResponse(Builder builder){
-            this.id = builder.id;
-            this.name = builder.name;
-            this.userId = builder.userId;
-            this.reward = builder.reward;
-            this.rewardInfoList = builder.rewardInfoList;
-        }
-
-        public static Builder builder(){
-            return new Builder();
-        }
-
-        public int getId(){
-            return id;
-        }
-
-        public String getName(){
-            return name;
-        }
-
-        public String getUserId(){
-            return userId;
-        }
-
-        public int getReward(){
-            return reward;
-        }
-
-        public List<RewardTagWrapperResponse> getRewardInfoList(){
-            return rewardInfoList;
-        }
-
-        public final static class Builder{
-
-            private int id;
-            private String name;
-            private String userId;
-            private int reward;
-            private List<RewardTagWrapperResponse> rewardInfoList;
-
-            private Builder(){}
-
-            public Builder id(int id){
-                this.id = id;
-                return this;
-            }
-
-            public Builder name(String name){
-                this.name = name;
-                return this;
-            }
-
-            public Builder userId(String userId){
-                this.userId = userId;
-                return this;
-            }
-
-            public Builder reward(int reward){
-                this.reward = reward;
-                return this;
-            }
-
-            public Builder rewardInfoList(List<RewardTagWrapperResponse> rewardInfoList){
-                this.rewardInfoList = rewardInfoList;
-                return this;
-            }
-
-            public UsersRewardTagResponse build(){
-                return new UsersRewardTagResponse(this);
-            }
-
-        }
-
-    }
-
-    public final static class RewardTagWrapperResponse{
-
-        @JsonProperty("reward_id")
-        private final int rewardId;
-
-        @JsonProperty("reward_tag_id")
-        private final int rewardTagId;
-
-        @JsonProperty("reward_tag_title")
-        private final String rewardTagTitle;
-
-        @JsonProperty("reward_presented_at")
-        private final String rewardPresentedAt;
-
-        @JsonProperty("reward_point")
-        private final int rewardPoint;
-
-        private RewardTagWrapperResponse(){
-            throw new UnsupportedOperationException("Can not invoke constructor \"RewardTagWrapperResponse()\"");
-        }
-
-        private RewardTagWrapperResponse(Builder builder){
-            this.rewardId = builder.rewardId;
-            this.rewardTagId = builder.rewardTagId;
-            this.rewardTagTitle = builder.rewardTagTitle;
-            this.rewardPresentedAt = builder.rewardPresentedAt;
-            this.rewardPoint = builder.rewardPoint;
-        }
-
-        public static Builder builder(){
-            return new Builder();
-        }
-
-        public int getRewardId(){
-            return rewardId;
-        }
-
-        public int getRewardTagId(){
-            return rewardTagId;
-        }
-
-        public String getRewardTagTitle(){
-            return rewardTagTitle;
-        }
-
-        public String getRewardPresentedAt(){
-            return rewardPresentedAt;
-        }
-
-        public int getRewardPoint(){
-            return rewardPoint;
-        }
-
-        public final static class Builder{
-
-            private int rewardId;
-            private int rewardTagId;
-            private String rewardTagTitle;
-            private String rewardPresentedAt;
-            private int rewardPoint;
-
-            private Builder(){}
-
-            public Builder rewardId(int rewardId){
-                this.rewardId = rewardId;
-                return this;
-            }
-
-            public Builder rewardTagId(int rewardTagId){
-                this.rewardTagId = rewardTagId;
-                return this;
-            }
-
-            public Builder rewardTagTitle(String rewardTagTitle){
-                this.rewardTagTitle = rewardTagTitle;
-                return this;
-            }
-
-            public Builder rewardPresentedAt(String rewardPresentedAt){
-                this.rewardPresentedAt = rewardPresentedAt;
-                return this;
-            }
-
-            public Builder rewardPoint(int rewardPoint){
-                this.rewardPoint = rewardPoint;
-                return this;
-            }
-
-            public RewardTagWrapperResponse build(){
-                return new RewardTagWrapperResponse(this);
-            }
-
-        }
-
     }
 
 }
