+++ waldreg/application-layer/reward-controller/src/main/java/org/waldreg/controller/reward/users/UsersRewardController.java
@@ -4,20 +4,28 @@ import java.util.ArrayList;
 import java.util.List;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 import org.wadlreg.reward.users.UsersRewardManager;
+import org.wadlreg.reward.users.dto.UsersRewardDto;
 import org.waldreg.character.aop.annotation.PermissionVerifying;
+import org.waldreg.controller.reward.users.mapper.ControllerUsersRewardMapper;
+import org.waldreg.controller.reward.users.response.RewardTagWrapperResponse;
+import org.waldreg.controller.reward.users.response.UsersRewardTagResponse;
 import org.waldreg.token.aop.annotation.Authenticating;
 
 @RestController
 public class UsersRewardController{
 
     private final UsersRewardManager usersRewardManager;
+    private final ControllerUsersRewardMapper controllerUsersRewardMapper;
 
     @Autowired
-    public UsersRewardController(UsersRewardManager usersRewardManager){
+    public UsersRewardController(UsersRewardManager usersRewardManager,
+            ControllerUsersRewardMapper controllerUsersRewardMapper){
         this.usersRewardManager = usersRewardManager;
+        this.controllerUsersRewardMapper = controllerUsersRewardMapper;
     }
 
     @Authenticating
@@ -35,6 +43,14 @@ public class UsersRewardController{
     @GetMapping("/reward-tag/reset")
     public void resetAllUsersReward(){
         usersRewardManager.resetAllUsersReward();
+    }
+
+    @Authenticating
+    @PermissionVerifying("Reward manager")
+    @GetMapping("/reward-tag/user/{id}")
+    public UsersRewardTagResponse readSpecifyUsersReward(@PathVariable("id") int id){
+        UsersRewardDto usersRewardDto = usersRewardManager.readSpecifyUsersReward(id);
+        return controllerUsersRewardMapper.usersRewardDtoToUsersRewardTagResponse(usersRewardDto);
     }
 
     private List<Integer> getUserIdList(String userId){
